+++ components/gfx/Cargo.toml
@@ -11,8 +11,14 @@ path = "lib.rs"
 [dependencies.plugins]
 path = "../plugins"
 
+[dependencies.gfx_traits]
+path = "../gfx_traits"
+
 [dependencies.net_traits]
 path = "../net_traits"
+
+[dependencies.canvas_traits]
+path = "../canvas_traits"
 
 [dependencies.util]
 path = "../util"
