+++ components/gfx/paint_context.rs
@@ -4,7 +4,7 @@
 
 //! Painting of display lists using Moz2D/Azure.
 
-use color;
+use gfx_traits::color;
 use display_list::TextOrientation::{SidewaysLeft, SidewaysRight, Upright};
 use display_list::{BLUR_INFLATION_FACTOR, BorderRadii, BoxShadowClipMode, ClippingRegion};
 use display_list::{TextDisplayItem};
