+++ components/layout/lib.rs
@@ -39,13 +39,14 @@ extern crate util;
 extern crate rustc_serialize;
 extern crate alloc;
 extern crate azure;
-extern crate canvas;
+extern crate canvas_traits;
 extern crate clock_ticks;
 extern crate collections;
 extern crate cssparser;
 extern crate encoding;
 extern crate geom;
 extern crate gfx;
+extern crate gfx_traits;
 extern crate layout_traits;
 extern crate libc;
 extern crate msg;
