+++ components/msg/Cargo.toml
@@ -1,5 +1,4 @@
 [package]
-
 name = "msg"
 version = "0.0.1"
 authors = ["The Servo Project Developers"]
