+++ components/script/lib.rs
@@ -40,6 +40,7 @@ extern crate png;
 extern crate rustc_serialize;
 extern crate time;
 extern crate canvas;
+extern crate canvas_traits;
 extern crate profile_traits;
 extern crate script_traits;
 extern crate selectors;
